fix(bvm): replace int64_t with int32_t to eliminate 64-bit GCC helpers on MVS
c2asm370 emits =A(@@MULDI3), =A(@@DIVDI3), =A(@@MODDI3), =A(@@NEGDI2),
=A(@@UMODDI), and =A(@@UDIVDI) for 64-bit arithmetic, but without EXTRN
declarations.  IFOX00 flags all of these as IFO188 (undefined symbol) and
returns RC=8, which failed the IRX#BVM build on MVS.

REXX NUMERIC DIGITS 9 caps integers at +/-999,999,999, which fits in
int32_t (max ~2.1B).  The integer fast-path cache, DO-loop counter, and
digit conversion helper are all safe with 32 bits.

Changed:
  - struct bc_stack_slot.int_cache: int64_t -> int32_t
  - struct vpool_entry.int_cache:   int64_t -> int32_t
  - struct bc_do_frame.counter:     int64_t -> int32_t
  - slot_to_int64 -> slot_to_int32 (return type + locals)
  - i64toa: int64_t/uint64_t -> int32_t/unsigned int
  - try_parse_int_cache: int64_t v -> int32_t v
  - try_arith_fast: int64_t va/vb/result -> int32_t
  - vpool_get_buf / vpool_set_buf signatures updated in .h and .c

Host tests: 36/36 tstbc_clean, full regression suite unaffected.

Fixes #147

+++ include/irxbvm.h
@@ -40,7 +40,7 @@ struct bc_stack_slot
 {
     PLstr str;          /* canonical string form; always present      */
     int32_t type_cache; /* 0=none, 1=integer, see LINTEGER_TY etc.   */
-    int64_t int_cache;  /* integer fast-path value                    */
+    int32_t int_cache;  /* integer fast-path value                    */
 };
 
 /* ================================================================== */
