+++ include/irxvpool.h
@@ -60,8 +60,8 @@ struct vpool_entry
     Lstr value;                      /* variable value              */
     int flags;                       /* VPOOL_DROPPED / _EXPOSED_REF */
     struct vpool_entry *exposed_ref; /* -> parent entry if exposed  */
-    int32_t type_cache; /* 0=unknown, IRXBC_STACK_LINTEGER=1        */
-    int64_t int_cache;  /* integer value when type_cache==1         */
+    int32_t type_cache;              /* 0=unknown, IRXBC_STACK_LINTEGER=1        */
+    int32_t int_cache;               /* integer value when type_cache==1         */
 };
 
 /* ================================================================== */
@@ -138,13 +138,13 @@ int vpool_get_buf(struct irx_vpool *pool,
                   const char *name_data, int name_len,
                   PLstr value,
                   int32_t *type_cache_out,
-                  int64_t *int_cache_out) asm("VPOOLGTB");
+                  int32_t *int_cache_out) asm("VPOOLGTB");
 
 int vpool_set_buf(struct irx_vpool *pool,
                   const char *name_data, int name_len,
                   const PLstr value,
                   int32_t type_cache,
-                  int64_t int_cache) asm("VPOOLSTB");
+                  int32_t int_cache) asm("VPOOLSTB");
 
 int vpool_drop_buf(struct irx_vpool *pool,
                    const char *name_data,
